+++ src/lib.rs
@@ -85,9 +85,10 @@ pub use crate::mkb::{
 pub use crate::mmc::{
     build_send_key_host_cert_chal, build_send_key_host_key, parse_media_id_response,
     parse_media_serial_response, parse_mkb_pack_response, parse_report_key_agid,
-    parse_report_key_drive_cert, parse_report_key_drive_cert_chal, parse_report_key_drive_key,
-    parse_send_key_host_cert_chal, parse_send_key_host_key, parse_volume_id_response, AgidResponse,
-    DataDirection, DriveCertChallengeResponse, DriveCertResponse, DriveCommand, DriveKeyResponse,
+    parse_report_key_binding_nonce, parse_report_key_drive_cert, parse_report_key_drive_cert_chal,
+    parse_report_key_drive_key, parse_send_key_host_cert_chal, parse_send_key_host_key,
+    parse_volume_id_response, AgidResponse, BindingNonceResponse, DataDirection,
+    DriveCertChallengeResponse, DriveCertResponse, DriveCommand, DriveKeyResponse,
     MediaIdentifierResponse, MediaSerialNumberResponse, MkbPackResponse, MockDrive,
     ReadDiscStructure, ReportKey, ScsiResponse, SendKey, VolumeIdResponse,
 };